7.3.7 GA8 release

+++ modules/apps/change-tracking/change-tracking-test/src/testIntegration/java/com/liferay/change/tracking/service/test/CTCollectionServiceTest.java
@@ -180,7 +180,7 @@ public class CTCollectionServiceTest {
 
 			Assert.assertTrue(resultSet.next());
 
-			Assert.assertEquals(0, resultSet.getLong(1));
+			Assert.assertEquals(0, resultSet.getInt(1));
 		}
 	}
